see changelog rev 1.78

+++ Musik/Classes/MusikPlayer.h
@@ -167,8 +167,8 @@ public:
 	void SetPlaylist		(const  CMusikSongArray &playlist ) {wxCriticalSectionLocker locker( m_critInternalData) ; m_Playlist = playlist;m_arrHistory.Clear(); }
 	void AddToPlaylist		( CMusikSongArray &songstoadd ,bool bPlayFirstAdded = true);	// NOTE this method, empties the songstoadd array.
 	void InsertToPlaylist(	 CMusikSongArray & songstoadd ,bool bPlayFirstInserted = true);  // NOTE this method, empties the songstoadd array.
-	void RemovePlaylistEntry( int index );
-	void MovePlaylistEntrys	(int nMoveTo ,const wxArrayInt &arrToMove);
+	void RemovePlaylistEntry( size_t index );
+	void MovePlaylistEntrys	(size_t nMoveTo ,const wxArrayInt &arrToMove);
 	void SetStartingNext	( bool bStart = true ){ m_StartingNext = bStart; }
 	void SetCrossfadeType	( int nType ){ m_CrossfadeType = nType; }
